update: delegate channel filtering to IFilterChannels
Took 24 minutes
