+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/channels/loadchannels/ILoadChannels.kt
@@ -1,0 +1,9 @@
+package io.getstream.chat.android.compose.viewmodel.channels.loadchannels
+
+import io.getstream.chat.android.client.api.models.QueryChannelsRequest
+
+internal interface ILoadChannels {
+    suspend fun load(queryChannelsRequest: QueryChannelsRequest)
+    suspend fun loadMore()
+}
+
