+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/channels/loadmessages/ILoadMessages.kt
@@ -1,0 +1,7 @@
+package io.getstream.chat.android.compose.viewmodel.channels.loadmessages
+
+internal interface ILoadMessages {
+    suspend fun load(query: String)
+    suspend fun loadMore()
+}
+
